requirements.txt
opencv-python-headless
onnxruntime
sherpa-onnx
sounddevice
requests
PyYAML
numpy
